+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-26 23:10:09
-// Commit: style(core): harmonize light and dark theme palette for core
+// Updated at: 2026-08-28 15:33:40
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '3964';
-  final String lastMessage = 'style(core): harmonize light and dark theme palette for core';
-  final DateTime timestamp = DateTime.parse('2026-08-26T23:10:09+05:00');
+  final String commitRef = '3992';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-08-28T15:33:40+05:00');
 }